requirements.txt
openai
tiktoken
google-generativeai
streamlit
python-dotenv
anthropic
curl_cffi
nltk